moves "createAddDependencies" to "createFractalProperties"
