--- src/index/formatFileStructure/toFractalTree/createAddDependency.ts
@@ -1,7 +1,0 @@
-export function createAddDependency(dependencyIndex: Record<string, string[]>) {
-  return (fileName: string, dependency: string) => {
-    const is = fileName in dependencyIndex;
-    if (!is) dependencyIndex[fileName] = [];
-    dependencyIndex[fileName].push(dependency);
-  };
-}
